Patch Daicon bug and switched to Rapier
As of writing this, Daicon has a bug where kinematic_daicon_node.gd is missing "if not d3: return" one two lines, fixing this restores compatibility with all physics engines and stops a random bug that occurs even with the default physics engine. Unfortunately Jolt has a bug where the player gets stuck on ramps, so we will stick with Rapier for now.